+++ src/vl-map-action.js
@@ -6,48 +6,22 @@ import {VlMap} from '/src/vl-map.js';
  * @class
  * @classdesc De abstracte kaart actie component.
  *
- * @property {boolean} active - Attribuut bepaalt of de kaart geactiveerd is.
  *
  * @extends HTMLElement
  * @mixes vlElement
  *
  * @property {boolean} data-vl-default-active - Attribuut wordt gebruikt om de actie standaard te activeren.
- * @property {boolean} data-vl-layer - Attribuut wordt gebruikt om via het naam attribuut de actie te koppelen aan een kaartlaag.
  *
  * @see {@link https://www.github.com/milieuinfo/webcomponent-vl-ui-map/releases/latest|Release notes}
  * @see {@link https://www.github.com/milieuinfo/webcomponent-vl-ui-map/issues|Issues}
  */
 export class VlMapAction extends vlElement(HTMLElement) {
-  static get _observedAttributes() {
-    return ['layer'];
-  }
-
   connectedCallback() {
-    this._layerChangedCallback();
     this.__registerMapActionChangedCallback();
   }
 
   static isVlMapAction() {
     return true;
-  }
-
-  /**
-   * Geeft de OL6 kaartlaag.
-   *
-   * @return {Object}
-   */
-  get layer() {
-    return this._layer;
-  }
-
-  /**
-   * Zet de kaartlaag.
-   *
-   * @param {Object} layer OL6 kaartlaag
-   */
-  set layer(layer) {
-    this._layer = layer;
-    this._processAction();
   }
 
   /**
@@ -63,11 +37,7 @@ export class VlMapAction extends vlElement(HTMLElement) {
     return this.closest('vl-map');
   }
 
-  get _layerElement() {
-    return this._mapElement.querySelector(`vl-map-layer[data-vl-name="${this.dataset.vlLayer}"]`) || this.closest('vl-map-layer');
-  }
-
-  get _active() {
+  get _defaultActive() {
     return this.hasAttribute('default-active');
   }
 
@@ -82,32 +52,17 @@ export class VlMapAction extends vlElement(HTMLElement) {
     this._mapElement.activateAction(this.action);
   }
 
-  _layerChangedCallback() {
-    if (this._layerElement) {
-      this.layer = this._layerElement.layer;
-    }
-  }
-
   _createAction() {
     console.warn('implementatie van _createAction ontbreekt');
   }
 
   _processAction() {
-    this._mapElement.ready.then(() => {
-      if (this._action) {
-        this._mapElement.removeAction(this._action);
+    if (this.action) {
+      this._mapElement.addAction(this.action);
+      if (this._defaultActive) {
+        this.activate();
       }
-
-      if (this.layer) {
-        this._action = this._createAction(this.layer);
-        if (this._action) {
-          this._mapElement.addAction(this._action);
-          if (this._active) {
-            this.activate();
-          }
-        }
-      }
-    });
+    }
   }
 
   __registerMapActionChangedCallback() {
