+++ src/vl-map-draw-point-action.js
@@ -1,4 +1,4 @@
-import {VlMapAction} from './vl-map-action.js';
+import {VlMapLayerAction} from './vl-map-layer-action.js';
 import {VlDrawAction, OlGeometryType} from '/node_modules/vl-mapactions/dist/vl-mapactions.js';
 
 /**
@@ -6,13 +6,13 @@ import {VlDrawAction, OlGeometryType} from '/node_modules/vl-mapactions/dist/vl-
  * @class
  * @classdesc De kaart teken actie component.
  *
- * @extends VlMapAction
+ * @extends VlMapLayerAction
  *
  * @see {@link https://www.github.com/milieuinfo/webcomponent-vl-ui-map/releases/latest|Release notes}
  * @see {@link https://www.github.com/milieuinfo/webcomponent-vl-ui-map/issues|Issues}
  * @see {@link https://webcomponenten.omgeving.vlaanderen.be/demo/vl-map-draw-actions.html|Demo}
  */
-export class VlMapDrawPointAction extends VlMapAction {
+export class VlMapDrawPointAction extends VlMapLayerAction {
   _createAction(layer) {
     return new VlDrawAction(layer, OlGeometryType.POINT, this._callback);
   }
